feat: publish MCP server
