+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "ui-kit.ai-monorepo",
-    "version": "0.0.6-alpha.12",
+    "version": "0.0.6-alpha.23",
     "private": true,
     "description": "",
     "homepage": "https://github.com/alex-mcgovern/ui-kit.ai#readme",
