+++ packages/mcp-server/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@ui-kit.ai/mcp-server",
-    "version": "0.0.6-alpha.12",
+    "version": "0.0.6-alpha.23",
     "repository": {
         "type": "git",
         "url": "https://github.com/alex-mcgovern/ui-kit.ai"
@@ -8,15 +8,21 @@
     "license": "MIT",
     "author": "Alex McGovern",
     "type": "module",
-    "bin": "./src/mcp-server/index.mjs",
+    "bin": "./src/index.mjs",
+    "files": [
+        "src"
+    ],
     "scripts": {
+        "check:eslint": "eslint .",
+        "check:prettier": "prettier . --check",
+        "check:vitest": "vitest run --passWithNoTests",
         "fix:eslint": "eslint . --fix",
-        "fix:prettier": "prettier . --write"
+        "fix:prettier": "prettier . --write",
+        "publish-packages": "pnpm publish --access public --no-git-checks"
     },
     "dependencies": {
         "@modelcontextprotocol/sdk": "1.9.0",
-        "react": "19.1.0",
-        "react-element-to-jsx-string": "17.0.1",
+        "@ui-kit.ai/metadata": "workspace:*",
         "zod": "3.24.1"
     },
     "devDependencies": {
@@ -24,9 +30,9 @@
         "@types/react": "19.1.3",
         "@ui-kit.ai/components": "workspace:*",
         "@ui-kit.ai/eslint": "workspace:*",
-        "@ui-kit.ai/metadata": "workspace:*",
         "@ui-kit.ai/prettier": "workspace:*",
         "@ui-kit.ai/tsconfig": "workspace:*",
+        "@ui-kit.ai/vite": "workspace:*",
         "@ui-kit.ai/vitest": "workspace:*"
     },
     "engines": {
