--- packages/mcp-server/tsconfig.json
@@ -1,10 +1,0 @@
-{
-    "extends": "@ui-kit.ai/tsconfig/react-library.json",
-    "include": [
-        "src/**/*",
-        "eslint.config.mjs",
-        ".prettierrc.mjs",
-        "../../scripts/utils/parse-story.ts"
-    ],
-    "exclude": ["node_modules"]
-}
