--- packages/metadata/vite.config.mjs
@@ -1,11 +1,0 @@
-import { buildViteConfig } from '@ui-kit.ai/vite'
-import path from 'node:path'
-
-import packageJson from './package.json' with { type: 'json' }
-
-export default buildViteConfig(
-    packageJson.name,
-    path.resolve(import.meta.dirname, 'tsconfig.json'),
-    [path.resolve(import.meta.dirname, 'src/index.ts')],
-    packageJson
-)
