+++ packages/storybook/src/scripts/gen-index.ts
@@ -29,20 +29,48 @@ async function main() {
         // Sort files alphabetically
         storyFiles.sort()
 
-        // Generate export statements
-        const exports = storyFiles.map((file) => {
+        // Start building the index file content
+        let content = `/**
+ * We expose all stories wrapped in \`composeStories\` which allows them to be consumed 
+ * as "portable" stories that can be used for testing and documentation purposes.
+ */
+
+import { composeStories as storybookComposeStories } from '@storybook/react'
+\n`
+
+        // Generate import statements for the stories
+        const imports = storyFiles.map((file) => {
             const baseName = path.basename(file, '.tsx')
             const componentName = kebabToPascalCase(baseName)
-            return `export * as ${componentName} from './stories/${baseName}'`
+            return `import * as ${componentName}Stories from './stories/${baseName}'`
         })
 
-        // Join exports with newlines
-        const content = exports.join('\n')
+        content += imports.join('\n') + '\n\n'
+
+        // Add compose function
+        content += `const compose = (
+    module: Parameters<typeof storybookComposeStories>[0]
+): ReturnType<typeof storybookComposeStories> => {
+    return storybookComposeStories(module, {
+        applyDecorators: (storyFn) => storyFn, // We don't want to apply any decorators
+    })
+}
+
+type Composed = ReturnType<typeof compose>\n\n`
+
+        // Generate the composed exports
+        const componentExports = storyFiles.map((file) => {
+            const baseName = path.basename(file, '.tsx')
+            const componentName = kebabToPascalCase(baseName)
+            return `export const ${componentName}: Composed = compose(${componentName}Stories)`
+        })
+
+        content += componentExports.join('\n')
 
         // Write to index.ts
         await fs.writeFile(indexPath, content)
 
-        console.info(`✅ Successfully generated ${indexPath}`)
+        console.info(`✅ Successfully generated ${indexPath} with ${storyFiles.length} components`)
     } catch (error) {
         console.error('❌ Error generating index.ts:', error)
         process.exit(1)
