+++ packages/tests/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@ui-kit.ai/tests",
-    "version": "0.0.6-alpha.12",
+    "version": "0.0.6-alpha.23",
     "private": true,
     "repository": {
         "type": "git",
